Passe la taille max d'un paquet de 2^16 octets à 2^32 octets

+++ src/client/client_c.h
@@ -53,7 +53,7 @@ private slots:
 private:
     //Socket
     QTcpSocket *m_socket;
-    quint16 m_taillePaquet;
+    quint32 m_taillePaquet;
 
     //Transfert
     QFile *m_fichier;
